requirements.txt
python-dotenv
requests
duckduckgo_search
flask # optional, for server functionality
waitress # optional, for server functionality